xspraydrift {XDriftAg}	R Documentation

Calculating spray-drift exposure with the XSprayDrift model

Description

xspraydrift calculates spray-drift exposure of receptors downwind of an exposure source. The function is based on the same measurement data as the rautmann90 function but uses fitted probability density functions for each distance with measurements. It can be used as a replacement for the rautmann90 function in scenarios where variability along the field-edge should be preserved.

Usage

xspraydrift(distance, q = stats::runif(1), target.exposure = NA_real_,
  crop = "arable", pdf.type = c("gamma", "normal", "lognormal", "beta",
  "uniform"), distance.bins = c("mean", "worst", "interpolated"),
  group = NULL)

Arguments

distance
A numeric vector of distances in meters. May contain NAs.

q
A numeric scalar specifying the quantile of the density function used for determining exposure. Must be in the interval [0; 1].

target.exposure
A numeric scalar specifying the exposure that is reported for distances of 0.

crop
Specifies the crop that is applied. See xspraydriftparameters for possible values.

pdf.type
Specifies the type of distribution to use. Must be "gamma", "normal", "lognormal", "beta" or "uniform".

distance.bins
Specifies the delineation method of distance classes. Must be "mean" for a delineation of classes in the middle between base measurement distances, "worst" for delineation at the measurement distances or "interpolated" to apply a linear regression between the nearest base mesurements.

group
Allows to specify an a priori group on which the distribution parameters are based. Possible values are "1-1-1", "1-1-2", "1-1-3", "1-1-4", "1-2-1", "1-2-2", "1-3-1", "2-1-1", "2-1-2", "2-1-4", "2-2-1", "2-2-2", "2-2-3" and "2-3-1", although not all groups are available for all crops. Refer to XDrift – An R Package to Simulate Spatially Explicit Pesticide Spray-drift Exposure of Non-Target-Species Habitats at Landscape Scales, Supplementary Online Material for details. If group is NULL, a lumped parameterization for the specified crop is used.

Details

The xspraydrift function works with a vector of distances in meters. Each distance is internally translated into a distance class according to the crop and distance.bins (see xspraydriftparameters for more details). Distances below the lower boundary of the first distance bin are considered to be actually a distance equal this lower boundary. Each distance class is associated with a parameter set for a distribution where the type of distribution is determined by the pdf.type argument (see also xspraydriftparameters) and, from this distribution, the quantile specified by q is sampled. The resulting value is returned as the fraction of exposure received by a sink located in the specified distance. For a distance of 0, the value of target.exposure is returned. A distance of NA results in a fraction of exposure of 0. This treatment of distances that are not positive numbers is in accordance with the topological functions of the XDrift package. By default, the parameters of the sampled distribution are derived from all experiments conducted on the specified crop using a lumped approach. By specifying the group argument, the distribution parameters can be restricted to an a priori defined group of experiments on the same crop, sharing similar spray pressure, wind speed and spraying technology. Please refer to XDrift – An R Package to Simulate Spatially Explicit Pesticide Spray-drift Exposure of Non-Target-Species Habitats at Landscape Scales, Supplementary Online Material for details.

Value

A vector of the length of distance that indicates the fraction of exposure at the given distances. Target receptors receive a value of target.exposure. Receptors with a distance of 0 result in a fraction of NA.

See Also

xspraydriftparameters for distribution parameters

Other spray-drift function: rautmann90

Examples

distances <- c(NA, 0:49)
xspraydrift(distances)

[Package XDriftAg version 1.0.0 Index]